:recycle: Split tests
