+++ tests/introspection/test_expression_processing.py
@@ -5,47 +5,6 @@ import pytest
 from json_logic.meta import JSONLogicExpression, Operation, Var
 
 
-def test_representation_simple_operation():
-    op1 = Operation("==", [10, "foo"])
-
-    expected = dedent(
-        """
-        Operation(==)
-          ├─ 10
-          └─ 'foo'
-    """
-    ).strip()
-    assert repr(op1) == expected
-
-
-def test_representation_nested_operations():
-    op1 = Operation("+", [10, 5])
-    op2 = Operation("==", [op1, 15])
-
-    expected = dedent(
-        """
-        Operation(==)
-          ├─ Operation(+)
-          │    ├─ 10
-          │    └─ 5
-          └─ 15
-    """
-    ).strip()
-
-    assert repr(op2) == expected
-
-
-def test_representation_var():
-    var_foo = Var("var", ["foo"])
-
-    assert repr(var_foo) == "$foo"
-
-
-def test_unknown_operator():
-    with pytest.raises(ValueError):
-        Operation("some-wordsalad that is highly unlikely to be an operation")
-
-
 @pytest.mark.parametrize(
     "expr",
     (None, "foo", 42, 0.05, True, False),
@@ -54,27 +13,21 @@ def test_expression_parser_primitives(expr):
     expression = JSONLogicExpression.from_expression(expr)
 
     assert expression.expression == expr
+    assert expression.as_tree() == expr
 
 
 def test_expression_parser_simple_operation():
     expression = JSONLogicExpression.from_expression({"var": ["foo"]})
 
     assert expression.expression == {"var": ["foo"]}
+    assert expression.as_tree() == Var("var", ["foo"])
 
 
 def test_expression_parser_simple_operation_syntactic_sugar():
     expression = JSONLogicExpression.from_expression({"var": "foo"})
 
     assert expression.expression == {"var": ["foo"]}
-
-
-def test_parse_simple_expression_into_tree():
-    expression = JSONLogicExpression.from_expression({"var": ["foo"]})
-
-    tree = expression.as_tree()
-
-    expected_tree = Var("var", ["foo"])
-    assert tree == expected_tree
+    assert expression.as_tree() == Var("var", ["foo"])
 
 
 def test_complex_expression_into_tree_with_representation():
